refactor(tests): fixed tsc errors by moving all *.ts files into src
